test: extract overlay interactions unit tests to own file
